backend/requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.7
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-base64==2.0
flake8==3.9.2
flake8-broken-line==0.3.0
flake8-isort==4.0.0
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
idna==3.2
ipdb==0.13.9
ipython==7.27.0
isort==5.9.3
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mccabe==0.6.1
oauthlib==3.1.1
parso==0.8.2
pep8-naming==0.12.1
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.20
psycopg2==2.9.1
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.10.0
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
reportlab==3.6.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
testfixtures==6.18.1
toml==0.10.2
traitlets==5.1.0
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
